extract source_ratings
