+++ linen-modern/src/main/scala/x7c1/linen/modern/init/dev/DummyFactory.scala
@@ -2,8 +2,8 @@ package x7c1.linen.modern.init.dev
 
 import android.content.Context
 import x7c1.linen.modern.accessor.DummyString.words
-import x7c1.linen.modern.accessor.database.{ChannelSourceMapParts, SourceParts}
-import x7c1.linen.modern.accessor.unread.{SourceRatingParts, SourceStatusAsStarted, SourceStatusParts}
+import x7c1.linen.modern.accessor.database.{SourceRatingParts, ChannelSourceMapParts, SourceParts}
+import x7c1.linen.modern.accessor.unread.{SourceStatusAsStarted, SourceStatusParts}
 import x7c1.linen.modern.accessor.{AccountAccessor, AccountParts, ChannelAccessor, ChannelParts, EntryParts, EntryUrl, LinenOpenHelper}
 import x7c1.linen.modern.struct.Date
 import x7c1.wheat.macros.logger.Log
